refactor(commands): separate move command text from move slash command
