+++ src/commands/move_thread/mod.rs
@@ -1,2 +1,3 @@
+pub mod common;
 pub mod slash_command;
 pub mod text_command;
