+++ src/commands/move_thread/slash_command/mod.rs
@@ -1,0 +1,1 @@
+pub mod move_thread;
