+++ src/errors/types.rs
@@ -184,7 +184,7 @@ impl fmt::Display for DiscordError {
             DiscordError::ChannelCreationFailed => write!(f, "Failed to create channel"),
             DiscordError::DmCreationFailed => write!(f, "Failed to create DM channel"),
             DiscordError::FailedToFetchCategories => write!(f, "Failed to fetch categories"),
-            DiscordError::FailedToMoveChannel => write!(f, "Failed to move channel"),
+            DiscordError::FailedToMoveChannel => write!(f, "Failed to move_thread channel"),
         }
     }
 }
